rename files to fit naming schema
